refactor: move FediAuthModal.tsx component
